requirements.txt
# Core Trading Dependencies
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Web Dashboard
Flask>=3.0.0
Flask-CORS>=4.0.0

# Database
sqlite3  # Built-in with Python

# Technical Analysis
ta-lib>=0.4.0  # Optional

# Data Visualization (Optional)
matplotlib>=3.7.0
plotly>=5.14.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Notifications (Optional)
# python-telegram-bot>=20.0
# requests>=2.28.0